added security to DELETE in api/trans/id

+++ src/app/api/transcriptions/[id]/route.js
@@ -7,9 +7,11 @@ import { auth } from '@clerk/nextjs/server'
 export async function GET(req, { params }) {
     const { userId } = await auth() //id from clerk
     const { id } = await params; //transctiption id from query params
+
     if (!userId) {
         return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
     }
+
     try {
         const user = await prisma.user.findUnique({ //user from db
             where: {
@@ -39,12 +41,35 @@ export async function GET(req, { params }) {
 }
 
 export async function DELETE(req, { params }) {
-    const { id } = await params;
+    const { userId } = await auth() //id from clerk
+    const { id } = await params; //transctiption id from query params
+
+    if (!userId) {
+        return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+    }
+
     try {
-        const transcription = await prisma.transcription.delete({
-            where: { id }
+        const user = await prisma.user.findUnique({ //user from db
+            where: {
+                clerkId: userId
+            }
         })
-        return NextResponse.json(transcription);
+        if (!user) return NextResponse.json({ error: "User not found" }, { status: 404 })
+
+        try {
+            const transcription = await prisma.transcription.delete({
+                where: { id }
+            })
+            if (!transcription) {
+                return NextResponse.json({ error: "Transcription to delete not found" }, { status: 404 });
+            }
+            if (transcription.userId !== user.id) {
+                return NextResponse.json({ error: 'Forbidden' }, { status: 403 })
+            }
+            return NextResponse.json(transcription);
+        } catch (error) {
+
+        }
     } catch (error) {
         return NextResponse.json({ error: "Error deleting transcription" }, { status: 500 })
     }
